Squashed 'lispBM/lispBM/' changes from 15fd3dea..02887bfd
02887bfd Change implementation of car family in evaluator for immediate termination on error
7517c2e1 bug fix in wait-for and addition of a few tests
275a8e46 removed pointless usage of WITH_GC
143d024c bumped version
a4836c7d closing a number of warnings when building 64bits. some still remain
e61c28e9 closing some warnings on 64bit build - lots to go
68a09916 a small amount of cleaning.
bd91e152 wait-for + trigger mechanism has been added, but is untested
70c82228 adding tests of undefine
003887d8 fix severe bug in undefine
bdc30576 Added two undefine tests
4e694bc8 experimental wake-up lbm task framework, work in progress
38068776 cleaning and refactoring
7e814674 cleaning and consolidating
b05615fa cleaning and consolidation of functionality
2b986f7c small refactor for
ce05734d removed dead code in evaluation of the progn continuation
9eeae272 made global env a static variable in env.c
1eacbf3f refactoring. may save a few cpu cycles in list allocation and GC

git-subtree-dir: lispBM/lispBM
git-subtree-split: 02887bfd12a8d98f2e6f73b3dbd38b9ec222358a

+++ src/extensions/array_extensions.c
@@ -913,7 +913,7 @@ lbm_value array_extension_buffer_length(lbm_value *args, lbm_uint argn) {
   return res;
 }
 
-
+//TODO: Have to think about 32 vs 64 bit here
 static lbm_value array_extensions_bufclear(lbm_value *args, lbm_uint argn) {
   lbm_value res = ENC_SYM_EERROR;
 
@@ -931,25 +931,25 @@ static lbm_value array_extensions_bufclear(lbm_value *args, lbm_uint argn) {
     clear_byte = (uint8_t)lbm_dec_as_u32(args[1]);
   }
 
-  unsigned int start = 0;
+  uint32_t start = 0;
   if (argn >= 3) {
     if (!lbm_is_number(args[2])) {
       return res;
     }
-    unsigned int start_new = lbm_dec_as_u32(args[2]);
+    uint32_t start_new = lbm_dec_as_u32(args[2]);
     if (start_new < array->size) {
       start = start_new;
     } else {
       return res;
     }
   }
-
-  unsigned int len = array->size - start;
+  // Truncates size on 64 bit build
+  uint32_t len = (uint32_t)array->size - start;
   if (argn >= 4) {
     if (!lbm_is_number(args[3])) {
       return res;
     }
-    unsigned int len_new = lbm_dec_as_u32(args[3]);
+    uint32_t len_new = lbm_dec_as_u32(args[3]);
     if (len_new <= len) {
       len = len_new;
     }
@@ -971,19 +971,19 @@ static lbm_value array_extensions_bufcpy(lbm_value *args, lbm_uint argn) {
 
   lbm_array_header_t *array1 = (lbm_array_header_t *)lbm_car(args[0]);
 
-  unsigned int start1 = lbm_dec_as_u32(args[1]);
+  uint32_t start1 = lbm_dec_as_u32(args[1]);
 
   lbm_array_header_t *array2 = (lbm_array_header_t *)lbm_car(args[2]);
 
-  unsigned int start2 = lbm_dec_as_u32(args[3]);
-  unsigned int len = lbm_dec_as_u32(args[4]);
+  uint32_t start2 = lbm_dec_as_u32(args[3]);
+  uint32_t len = lbm_dec_as_u32(args[4]);
 
   if (start1 < array1->size && start2 < array2->size) {
     if (len > (array1->size - start1)) {
-      len = (array1->size - start1);
+      len = ((uint32_t)array1->size - start1);
     }
     if (len > (array2->size - start2)) {
-      len = (array2->size - start2);
+      len = ((uint32_t)array2->size - start2);
     }
 
     memcpy((char*)array1->data + start1, (char*)array2->data + start2, len);
